Fragment view switching is now working.
