+++ LocationPrivacyApp/app/src/main/java/com/djsg38/locationprivacyapp/PreferenceList.java
@@ -1,31 +1,17 @@
 package com.djsg38.locationprivacyapp;
 
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.app.FragmentTransaction;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.LinearLayout;
 
-import com.djsg38.locationprivacyapp.models.Preference;
-import com.djsg38.locationprivacyapp.models.Session;
-
-import java.util.Random;
 
 import io.realm.Realm;
-import io.realm.RealmChangeListener;
 
 public class PreferenceList
         extends AppCompatActivity
         implements PreferenceFragment.OnFragmentInteractionListener {
 
     Realm realm;
-    Random rand;
 
     public void onFragmentInteraction(int position) {
         return;
@@ -39,41 +25,12 @@ public class PreferenceList
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        getFragmentManager()
-                .beginTransaction()
-                .add(R.id.fragment_content, new PreferenceListFragment())
-                .commit();
-
         realm = Realm.getDefaultInstance();
 
-        rand = new Random();
-
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.floatingActionButton);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(final View view) {
-                realm.executeTransactionAsync(new Realm.Transaction() {
-                    @Override
-                    public void execute(Realm realm) {
-                        Session session = realm.where(Session.class).findFirst();
-                        Preference preference = realm.createObject(Preference.class, "Doot" + String.valueOf(rand.nextInt() % 100));
-                        preference.setPrivacyScale(1.0);
-                        preference.setService(false);
-                        session.getPreferences().add(preference);
-                    }
-                }, new Realm.Transaction.OnSuccess() {
-                    @Override
-                    public void onSuccess() {
-                        Snackbar.make(view, "successfully added new pref", 2000).show();
-                    }
-                }, new Realm.Transaction.OnError() {
-                    @Override
-                    public void onError(Throwable error) {
-                        error.printStackTrace();
-                    }
-                });
-            }
-        });
+        getFragmentManager()
+                .beginTransaction()
+                .add(R.id.content_preference_list, new PreferenceListFragment())
+                .commit();
 
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
     }
