+++ LocationPrivacyApp/app/src/main/res/layout/activity_preference_list.xml
@@ -2,6 +2,7 @@
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/activity_preference_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
@@ -22,11 +23,6 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/preference_list_container">
-
-    </FrameLayout>
+    <include layout="@layout/content_preference_list"/>
 
 </android.support.design.widget.CoordinatorLayout>
