+++ LocationPrivacyApp/app/src/main/res/layout/content_preference_list.xml
@@ -1,9 +1,11 @@
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_content"
+    android:id="@+id/content_preference_list"
     android:name="com.djsg38.locationprivacyapp.PreferenceListFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:layout="@layout/fragment_preference_list" />
+    tools:showIn="@layout/activity_preference_list">
+
+</FrameLayout>
